"Add event ticketing functionality to Google Wallet gem"
This commit introduces a significant expansion of the Google Wallet gem's capabilities by adding feature-rich support for event ticketing operations.

- Removed Travis CI configs, as they were no longer needed
- Added resource and operation classes for Event Tickets, providing methods for creating, checking the existence of, and updating ticket classes and objects
- Introduced new Google Wallet configuration options and authentication handling, allowing the user to load JSON credentials from a file and fetch access tokens
- Added HTTParty and JWT as dependencies to handle HTTP requests and JSON Web tokens respectively
- Updated the GoogleWallet to glue together all the new classes and operations.

The primary motivation behind these changes was to enrich the Google Wallet gem with new features that align it more closely with the diverse needs of real-world applications, such as handling event ticketing. These updates make it possible to integrate Google Wallet into applications that deal with event management in a more seamless and efficient manner.